MAP-2010 add sortable-search-results to front end js
