+++ server/views/pages/search-for-prisoner.njk
@@ -158,8 +158,3 @@
 </div>
 
 {% endblock %}
-
-
-{% block script %}
-<script src="/assets/sortable-search-results.js"></script>
-{% endblock %}
